+++ client/src/app/modules/transactions/components/transaction-list/transaction-list.component.spec.ts
@@ -5,7 +5,7 @@ import { By } from '@angular/platform-browser';
 import { DebugElement } from '@angular/core';
 import { MaterialModule } from '@angular/material';
 import { RouterTestingModule } from '@angular/router/testing';
-import { reducer } from '../../reducers';
+import { reducer } from '../../../../reducers';
 import { StoreModule } from '@ngrx/store';
 
 import { TransactionListComponent } from './transaction-list.component';
