+++ client/src/app/modules/transactions/transactions.module.ts
@@ -14,8 +14,6 @@ import { TransactionListComponent } from './components/transaction-list/transact
 import { TransactionComponent } from './components/transaction/transaction.component';
 import { TransactionDetailComponent } from './components/transaction-detail/transaction-detail.component';
 
-import { reducer } from './reducers';
-
 import { TransactionEffects } from './effects/transaction';
 
 import { TransactionService } from './services/transaction.service';
@@ -32,7 +30,6 @@ import { TransactionService } from './services/transaction.service';
     HttpModule,
     TransactionsRoutingModule,
     MaterialModule,
-    StoreModule.provideStore(reducer),
     StoreDevtoolsModule.instrumentOnlyWithExtension(),
     EffectsModule.run(TransactionEffects),
     AgmCoreModule.forRoot({
