+++ client/src/app/reducers/index.ts
@@ -37,6 +37,8 @@ import { combineReducers } from '@ngrx/store';
  * notation packages up all of the exports into a single object.
  */
 import * as fromUser from './user';
+import * as fromTransaction from '../modules/transactions/reducers/transaction';
+import * as fromTransactionMonth from '../modules/transactions/reducers/transaction-month';
 
 
 /**
@@ -45,6 +47,8 @@ import * as fromUser from './user';
  */
 export interface State {
   user: fromUser.State;
+  transaction: fromTransaction.State;
+  transactionMonth: fromTransactionMonth.State;
 }
 
 
@@ -57,6 +61,8 @@ export interface State {
  */
 const reducers = {
   user: fromUser.reducer,
+  transaction: fromTransaction.reducer,
+  transactionMonth: fromTransactionMonth.reducer,
 };
 
 const developmentReducer: ActionReducer<State> = compose(storeFreeze, combineReducers)(reducers);
@@ -87,3 +93,5 @@ export function reducer(state: any, action: any) {
  * ```
  */
 export const getUserState = (state: State) => state.user;
+export const getTransactionState = (state: State) => state.transaction;
+export const getTransactionMonthState = (state: State) => state.transactionMonth;
